+++ docs/helpers/PrimaryControls.jsx
@@ -3,7 +3,7 @@ import { Search } from 'reactabular';
 import PerPage from './PerPage';
 
 const PrimaryControls = ({
-  perPage, columns, rows,
+  perPage, columns, rows, query,
   onPerPage, onSearch,
   ...props
 }) => (
@@ -12,7 +12,7 @@ const PrimaryControls = ({
       <PerPage value={perPage} onChange={onPerPage} />
     </div>
     <div className="search-container">
-      Search <Search columns={columns} rows={rows} onChange={onSearch} />
+      Search <Search query={query} columns={columns} rows={rows} onChange={onSearch} />
     </div>
   </div>
 );
@@ -20,6 +20,7 @@ PrimaryControls.propTypes = {
   perPage: React.PropTypes.number,
   columns: React.PropTypes.array,
   rows: React.PropTypes.array,
+  query: React.PropTypes.object,
   onPerPage: React.PropTypes.func,
   onSearch: React.PropTypes.func
 };