+++ packages/reactabular-search-field/CHANGELOG.md
@@ -1,3 +1,8 @@
+1.2.5 / 2016-08-08
+==================
+
+  * Improvement - Accept `query` as a prop now. Better for persistency.
+
 1.2.3 / 2016-08-08
 ==================